Extract CampaignDetailsHeader to own component
